+++ app/pages/tne.vue
@@ -267,6 +267,7 @@ import { computed, shallowRef, watch } from 'vue'
 import { useQuery, useLazyQuery } from '@vue/apollo-composable'
 import { useFlexAreaFormatting } from '~/composables/useFlexAreaFormatting'
 import { useBufferDetails } from '~/composables/useBufferDetails'
+import { useCensusGeographyLayers } from '~/composables/useCensusGeographyLayers'
 import {
   getFlexAreaType,
   getFlexAdvanceNotice,
@@ -611,66 +612,7 @@ const flexFeaturesForReport = computed((): Feature[] => {
 // Geography datasets
 /////////////////
 
-const {
-  result: censusGeographyResult,
-} = useQuery<{ census_datasets: CensusDataset[] }>(
-  geographyLayerQuery,
-  () => ({
-    dataset_name: geoDatasetName.value,
-    geography_ids: geographyIds.value,
-    include_geographies: geographyIds.value.length > 0,
-  })
-)
-
-const CENSUS_LAYER_DISPLAY_NAMES: Record<string, string> = {
-  place: 'City / Place',
-  county: 'County',
-  state: 'State',
-  tract: 'Census Tract',
-  bg: 'Block Group',
-  blockgroup: 'Block Group',
-  tabblock20: 'Census Block (2020)',
-  county_subdivision: 'County Subdivision',
-  urban_area: 'Urban Area',
-  uac20: 'Urban Area (2020)',
-  cbsa: 'Core-Based Statistical Area (CBSA)',
-  csa: 'Combined Statistical Area (CSA)',
-  zip: 'ZIP Code (ZCTA)',
-  zcta: 'ZIP Code (ZCTA)',
-  zcta520: 'ZIP Code (ZCTA, 2020)',
-  cd: 'Congressional District',
-  cd119: 'Congressional District (119th)',
-  congressional_district: 'Congressional District',
-  sldu: 'State Legislative District (Upper)',
-  sldl: 'State Legislative District (Lower)',
-}
-
-function formatCensusLayerLabel (_dsName: string, layerName: string): string {
-  // Use human-readable name when known; otherwise strip "Layer: " prefix and title-case
-  return CENSUS_LAYER_DISPLAY_NAMES[layerName]
-    ?? layerName.replace(/^layer:\s*/i, '').replace(/_/g, ' ').replace(/\b\w/g, c => c.toUpperCase())
-}
-
-const censusGeographyLayerOptions = computed(() => {
-  const geomDatasets = censusGeographyResult.value?.census_datasets || []
-  const options = []
-  for (const ds of geomDatasets || []) {
-    for (const layer of ds.layers || []) {
-      options.push({ value: layer.name, label: formatCensusLayerLabel(ds.name, layer.name) })
-    }
-  }
-  return options
-})
-
-const censusGeographiesSelected = computed((): CensusGeography[] => {
-  const ret: CensusGeography[] = []
-  for (const ds of censusGeographyResult.value?.census_datasets || []) {
-    for (const geo of ds.geographies || []) {
-      ret.push(geo)
-    }
-  }
-  return ret
-})
+const { censusGeographyLayerOptions, censusGeographiesSelected } = useCensusGeographyLayers()
 
 const acsDatasetLabel = computed(() => formatAcsDatasetLabel(SCENARIO_DEFAULTS.tableDatasetName))
