+++ src/scenario/index.ts
@@ -5,3 +5,4 @@ export * from './buffer-passes'
 export * from './stop-clusters'
 export * from './phases'
 export * from './report-tables'
+export * from './buffer-details'
